task completed and finaly optimise code
